Keep the range analysis warning-neutral and give the interp its Range index
Splits the #1400 analysis into its own include so mod_p4.rs stays under the
800-line codopsy max-lines threshold, requires at least one for-in head read
so a bind with no reader still materializes (the #1272 pin), moves the
4-billion-iteration cell to its own fixture so its unavoidable interp
abstention does not cost the guard rails their third vote, and teaches the
interpreter to index a Range instead of dissenting from a correct consensus.
