+++ crates/almide-interp/interp-abstain-ledger.txt
@@ -103,6 +103,7 @@ place_mutation  out-of-interp-scope capability: prim.alloc_map
 pure_bang_propagation  out-of-interp-scope capability: prim.handle
 r5_lowmisc_param_try_err  out-of-interp-scope capability: prim.alloc_map
 random_int_entropy  out-of-interp-scope capability: prim.alloc_bytes
+range_bind_huge  interp fuel/recursion budget exhausted
 rc_alloc_stress  out-of-interp-scope capability: prim.alloc_map
 regex_engine  out-of-interp-scope capability: prim.handle
 regex_fuzz_batch  out-of-interp-scope capability: prim.handle
