+++ docs/contracts/conformance.md
@@ -11,7 +11,7 @@
 > (spec-coverage + evidence-class >= fixture for every active contract), so this
 > page cannot legitimately contain an empty Fixtures cell.
 
-97 normative sections; 437 distinct executable fixtures.
+97 normative sections; 438 distinct executable fixtures.
 
 | Section | Contracts | Fixtures (how CI runs each) |
 |---------|-----------|------------------------------|
@@ -41,7 +41,7 @@
 | ALS-E7 | C-235 | `spec/wasm_cross/grouping_unary.almd` (byte-compare) |
 | ALS-E8 | C-236 | `spec/wasm_cross/tuple_ops.almd` (byte-compare) |
 | ALS-E9 | C-237 | `spec/wasm_cross/option_result_ctors.almd` (byte-compare) |
-| ALS-E10 | C-238 | `spec/wasm_cross/range_first_class.almd` (byte-compare)<br>`spec/wasm_cross/range_bind_no_materialize.almd` (byte-compare) |
+| ALS-E10 | C-238 | `spec/wasm_cross/range_first_class.almd` (byte-compare)<br>`spec/wasm_cross/range_bind_no_materialize.almd` (byte-compare)<br>`spec/wasm_cross/range_bind_huge.almd` (byte-compare) |
 | ALS-E11 | C-239 | `spec/wasm_cross/collection_literals.almd` (byte-compare) |
 | ALS-E12 | C-240 | `spec/wasm_cross/collection_literals.almd` (byte-compare) |
 | ALS-E13 | C-242 | `spec/wasm_cross/if_block_forms.almd` (byte-compare) |
