+++ spec/wasm_cross/range_bind_huge.almd
@@ -1,0 +1,22 @@
+// @contract: C-238
+// #1400, the load-bearing cell, ALONE in its own file because it is the one cell
+// the reference interpreter cannot vote on: 4294967294 iterations exhaust the
+// tree-walker's fuel budget by construction, and a whole-file abstention would
+// cost the sibling guard rails in range_bind_no_materialize.almd their third
+// vote too. The abstention is ledgered for THIS file only.
+//
+// A `let`-bound range used only as a `for-in` head is not materialized. A
+// materializing leg needs ~32 GB for the block; a counting leg needs two locals.
+// Before #1400 the wasm leg materialized (#1272's fix) while native iterated
+// lazily, so this printed the sum natively and died `Error: out of memory` on
+// wasm — found by the 2026-08-13 fuzz-nightly (run 31667874852, seed
+// 506685997633 index 522). 0 + 1 + ... + 4294967293 = 9223372030412324865,
+// which is representable in i64 (it is 942 short of 2^63-1).
+effect fn main() -> Unit = {
+  let huge = 0..<4294967295
+  var hs = 0
+  for i in huge {
+    hs = hs + i
+  }
+  println("huge=" + int.to_string(hs))
+}
